@@ -1,30 +1,26 @@
 name: video_editor_example
 description: The example project for the video_editor package.
-version: 0.0.1
 publish_to: "none"
+version: 0.0.1
 
 environment:
-  sdk: ">=3.0.0 <4.0.0"
+  sdk: ^3.5.1
 
 dependencies:
   flutter:
     sdk: flutter
   video_editor_2:
     path: ../
-  ffmpeg_kit_flutter_min_gpl: ^5.1.0
-  fraction: ^4.1.4
-  helpers: ^1.1.4
-  image_picker: ^0.8.6+1
-  path: ^1.8.2
-  video_player: ^2.5.1
-  path_provider: ^2.0.12
-  cross_file: ^0.3.3+4
-  ffmpeg_wasm: ^1.0.1
+  fraction: ^5.0.3
+  image_picker: ^1.1.2
+  video_player: ^2.9.1
+  path_provider: ^2.1.4
+  path: ^1.9.0
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  flutter_lints: ^2.0.1
+  flutter_lints: ^4.0.0
 
 flutter:
   uses-material-design: true
